Allow switching to different qt versions easily
We can defining QT_MAJOR_VERSION when calling
cmake to specify which qt version we want to
use.

We use QT5 by default as the latest KDE stable
version is KDE5.

We require QT 5.15 because we are using
versionless targets.
See https://doc.qt.io/qt-6/cmake-qt5-and-qt6-compatibility.html
for more info.